+++ src/sdk/CHANGELOG.md
@@ -20,6 +20,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - RateLimit header processing capability, turned off by default [jansenbe - Bert Jansen]
 - Added `DeleteRecycledSiteCollection` methods to delete site collections from the recycle bin [jansenbe - Bert Jansen]
 - Admin library: ability to specify ResourceBehaviorOptions when creating a Group using application permissions #978 [koenzomers - Koen Zomers]
+- Admin library: Ability to specify members on team site creation time #981 [LeHailender - Jens Haile]
 
 ### Changed